Added tests for PlayerService and erased recursive constraint in Game class
